// src/app/shared/models/Dtos/eventoDto.ts
export interface EventoDto{
    codigoevento: number;
    codigomotivoevento: number;
    codigorecepcionevento: number;
    codigopuntoventa: number;
    cafccompraventa: string;
    cafctelecom: string;
    cuis: string;
    cufd: string;
    cufdevento: string;
    descripcion: string;
    fechahorainicioevento: Date;
    fechafinevento: Date;
    activo: boolean
}